Fix eb build

+++ libobs-d3d11/d3d11-shader.cpp
@@ -257,11 +257,11 @@ void gs_shader::Compile(const char *shaderString, const char *file,
 			cacheFile.seekg(0, ios::beg);
 
 			len -= sizeof(checksum);
-			device->d3dCreateBlob(len, shader);
+			device->d3dCreateBlob((SIZE_T)len, shader);
 			cacheFile.read((char *)(*shader)->GetBufferPointer(),
 				       len);
 			uint64_t calculated_checksum = fnv1a_hash(
-				(char *)(*shader)->GetBufferPointer(), len);
+				(char *)(*shader)->GetBufferPointer(), (size_t)len);
 
 			cacheFile.read((char *)&checksum, sizeof(checksum));
 			if (calculated_checksum != checksum)
